feat: :sparkles: ajuste no banco e insercao de dados para test

+++ database/models/product.js
@@ -7,7 +7,7 @@ const Product = (sequelize, DataTypes) => {
       type: DataTypes.INTEGER
     },
     name: DataTypes.STRING,
-    price: DataTypes.INTEGER,
+    price: DataTypes.DOUBLE,
     quantity: DataTypes.INTEGER
   }, {
     tableName: 'Products',
